Merged branch "common_dataset_menu_object".
Working WidgetContainer.dataSetMenu as a DataSetMenu instance.

This necessitated a special hook in RasterWidget extensions to reset the
dataset menu (set it's menu version to -1) when the dataset is changed locally.
This is due to the timing of invoking the widget UpdateState() method in a
wx.CallAfter(), which means the widget updates its curDataSet after a menu
UpdateMenu() call has already incremented the menu version.