WIP, addressed :node datasets in CoreAxial2DView and TimePlots.
Pushed out as version 1.0.64.

Updated pin_averages.py and channel_averages.py, hopefully fixing a bug in the
weights on the former.  Fixed the node display axis for 'xz' versus 'yz' in
CoreAxial2DView.  Refactored it to have single _CreateDrawConfig() and
_CreateRasterImage() methods that internally handle self.nodalMode.

Also fixed a bug where toolbar item event handlers were not removed upon load
of a new file, resulting in multiple widgets when a toolbar was clicked.