Averages should be working again.
Removed all the mess to save factors for a dataset in the quick-and-dirty
effort to visualize weights.  The duh-for-me deal was not recognizing
np.ndarrays are iterable, so calling hasattr( result, '__iter__' ) to determine
if a tuple (averaged data, factors) is returned channel_averages or a just the
averaged data from pin_averages doesn't work.